test: 테스트 코드 정리한다
 - FilteringSlowTest, LargeCatSlowTest, SmallCatMapperTest 모두 필터링
   관련 기능을 테스트하고 있어서 한 군데로 모았다
