+++ src/test/java/org/swyp/weddy/slow/FilteringSlowTest.java
@@ -12,7 +12,11 @@ import org.swyp.weddy.domain.checklist.dao.LargeCatMapper;
 import org.swyp.weddy.domain.checklist.entity.LargeCatItem;
 import org.swyp.weddy.domain.checklist.exception.LargeCatItemNotExistsException;
 import org.swyp.weddy.domain.checklist.service.FilteringService;
+import org.swyp.weddy.domain.checklist.service.SmallCatService;
 import org.swyp.weddy.domain.checklist.service.dto.FilterByStatusDto;
+import org.swyp.weddy.domain.checklist.service.dto.SmallCatItemSelectDto;
+import org.swyp.weddy.domain.checklist.web.response.LargeCatItemResponse;
+import org.swyp.weddy.domain.checklist.web.response.SmallCatItemPreviewResponse;
 
 import java.util.List;
 
@@ -22,6 +26,8 @@ import java.util.List;
 public class FilteringSlowTest {
     @Autowired
     private FilteringService filteringService;
+    @Autowired
+    private SmallCatService smallCatService;
     @Autowired
     private LargeCatMapper largeCatMapper;
     @Autowired
@@ -50,5 +56,67 @@ public class FilteringSlowTest {
         org.junit.jupiter.api.Assertions.assertThrows(LargeCatItemNotExistsException.class, () -> {
             filteringService.filterByStatus(new FilterByStatusDto(1L, List.of("시작전")));
         });
+    }
+
+    @DisplayName("진행 상황 기준으로 대분류 항목을 필터링할 수 있다")
+    @Test
+    void filter_large_cat_items_by_status() {
+        jdbcTemplate.update("""
+                update small_category_item
+                set status_id = 1
+                where title in ('스튜디오 / 스냅 예약' , '드레스투어 일정 예약' , '스튜디오 드레스 선택' , '본식 드레스 선택')
+                and small_category_item.large_category_item_id = 3;
+                """);
+
+        jdbcTemplate.update("""
+                update small_category_item
+                set status_id = 1
+                where title in ('신혼집 계약', '가전/가구 구매')
+                and small_category_item.large_category_item_id = 8;
+                        """);
+
+        jdbcTemplate.update("""
+                update small_category_item
+                set status_id = 2
+                where title in ('침구/생활용품 구매', '주방용품 구매', '신혼집 입주')
+                and small_category_item.large_category_item_id = 8;
+                """);
+
+        List<LargeCatItemResponse> res = filteringService.filterByStatus(new FilterByStatusDto(1L, List.of("시작전")));
+        List<LargeCatItemResponse> res2 = filteringService.filterByStatus(new FilterByStatusDto(1L, List.of("시작전", "진행중")));
+
+        Assertions.assertThat(res).isNotNull();
+        Assertions.assertThat(res2).isNotNull();
+    }
+
+    @DisplayName("진행 상황 기준으로 소분류 항목을 필터링할 수 있다")
+    @Test
+    void filter_small_cat_items_by_status() {
+
+        jdbcTemplate.update("""
+                update small_category_item
+                set status_id = 1
+                where title in ('스튜디오 / 스냅 예약' , '드레스투어 일정 예약' , '스튜디오 드레스 선택' , '본식 드레스 선택')
+                and small_category_item.large_category_item_id = 3;
+                """);
+
+        jdbcTemplate.update("""
+                update small_category_item
+                set status_id = 1
+                where title in ('신혼집 계약', '가전/가구 구매')
+                and small_category_item.large_category_item_id = 8;
+                        """);
+
+        jdbcTemplate.update("""
+                update small_category_item
+                set status_id = 2
+                where title in ('침구/생활용품 구매', '주방용품 구매', '신혼집 입주')
+                and small_category_item.large_category_item_id = 8;
+                """);
+
+        List<SmallCatItemPreviewResponse> res = smallCatService.findItemPreviewsByStatus(new SmallCatItemSelectDto(1L, 3L, List.of("시작전")));
+        List<SmallCatItemPreviewResponse> res2 = smallCatService.findItemPreviewsByStatus(new SmallCatItemSelectDto(1L, 8L, List.of("시작전", "진행중")));
+        Assertions.assertThat(res).isNotNull();
+        Assertions.assertThat(res2).isNotNull();
     }
 }
